Scrapped flag(portable) in favor of os(windows)
The portable implementation only works on Windows because it relies on C
implementations for some features. I think it's better we spell that out
rather than tip toe around it.

The UNIX_IMPL define was also removed because mingw32_HOST_OS is already
defined for Windows systems. No need to complicate the cabal file with
information that is already available.