Removed the old testsuite as it didn't give us much
The plan was for unix-compat to be testable on a posix platform by
comparing the portable implementation to the unix package. Unfortunately
the portable implementation contains too much Windows specific code for
this to be possible.

Because of the nature of the functions which unix-compat implements,
we'll just have to rely on manual testing for now. In time I might
consider creating some tests which exercise the code in a different way.