Fix compatibilities with unix >=2.8 package
Since version 2.8, the unix package changed `UserEntry` and `GroupEntry`
into the `ByteString` based implementation instead of `String` as a
breaking change. We can compile this package with unix >=2.8 if just
re-exports the two of new data types, but it's probably uncomfortable
for users because types of fields of the two isn't the same between
prior 2.8 and since 2.8. So we need to define and export compatible data
types for `UserEntry` and `GroupEntry` to resolve this problem.